requirements.txt
streamlit
faiss-cpu
sentence-transformers
numpy
google-genai
pypdf
requests==2.32.5
